Delete modeltraining.ipynb
I’ve censored some photos in the PDF for privacy reasons. The file training_ipynb.pdf is a replacement of modeltraining and includes the relevant code and accuracy details — please refer to it instead. If you’d like the complete .ipynb file, feel free to email me at parthgupta9999@gmail.com — I’ll be happy to share it with you.